refactor(nudges): extract persistence helpers to fix max-lines lint error
